\name{key<-}
\alias{key<-}
\title{ Deprecated. }
\keyword{internal}
\description{
  This function is deprecated. It will be removed in future. Please use \code{\link{setkey}}.
}
\usage{
key(x) <- value   #  warning since 2012; DEPRECATED since Mar 2019
}
\examples{
# dummy example section to pass release check that all .Rd files have examples
}
\arguments{
\item{x}{ Deprecated. }
}
