\name{dt}
\alias{as.data.table}
\alias{as.data.frame.data.table}
\alias{as.data.table.data.frame}
\alias{as.data.table.data.table}
\alias{as.data.table.matrix}
\alias{as.matrix.data.table}
\alias{cbind.data.table}
\alias{[<-.data.table}
\alias{[.data.table}
\alias{dim.data.table}
\alias{dimnames<-.data.table}
\alias{dimnames.data.table}
\alias{head.data.table}
\alias{is.data.table}
\alias{NCOL}
\alias{NROW}
\alias{print.data.table}
\alias{rbind.data.table}
\alias{subset.data.table} 
\alias{tail.data.table}
\title{ data.table support functions }
\description{
  These are largely direct copies of the base functions for data.frame with the code manipulating row.names removed.
}
\usage{
}
\arguments{
}
\details{
  These functions work exactly as their data.frame counterparts, subject to the differences between data.frame and data.table documented under help(data.table).
}
\value{
}
\references{ }
\author{ Matt Dowle }
\note{ }
\seealso{ \code{\link{data.table}}, \code{\link{tables}}, \code{\link{data.frame}} }
\examples{
}
\keyword{ data }


