\name{set2key}
\alias{set2key}
\alias{set2keyv}
\alias{key2}
\alias{key<-}
\title{ Deprecated. }
\description{
  These functions are deprecated. They will be removed in future. Please use the functions in \code{\link{setkey}}.
}
\usage{
set2key(...)      #  DEPRECATED; helpful error since May 2018 and warning since Nov 2016
set2keyv(...)     #  DEPRECATED; helpful error since May 2018 and warning since Nov 2016
key2(...)         #  DEPRECATED; helpful error since May 2018 and warning since Nov 2016
key(x) <- value   #  warning since 2012; DEPRECATED since Mar 2019
}
\arguments{
\item{\dots}{ Deprecated. }
\item{x}{ Deprecated. }
\item{value}{ Deprecated. }
}

