\name{coalesce}
\alias{coalesce}
\title{ Coalescing missing values }
\description{
    Fill in missing values in a vector by successively pulling from candidate vectors in order (\emph{a la} the ANSI SQL function of the same name).

}
\usage{
  coalesce(x, \dots, .dots=NULL)
}
\arguments{
  \item{x}{ A vector of values }
  \item{\dots}{ Vectors to be used successively as replacements for missing values of \code{x}; singletons are also accepted. }
  \item{.dots}{ Alternative interface to specify replacements by supplying a list directly. }
}
\details{

}

\value{

}
\examples{

}
\keyword{ data }


