\name{last}
\alias{last}
\title{ Last item of an object }
\description{
   Returns last item of a vector, or last row of a matrix, data.frame or data.table.
}
\usage{
last(x)
}
\arguments{
  \item{x}{ A vector, matrix, data.frame or data.table }
}
% \details{
% }
\value{
    The last item. If x is a data.table, the last row is returned at a one row data.table.
}
% \references{ }
\author{ Matt Dowle }
% \note{ }
\seealso{ \code{\link{NROW}} }
% \examples{
% }
\keyword{ data }

